# Pyt_volshebnika/packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wizard Path Game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthRequest {
  name: string;
  password: string;
}

export interface AuthResponse {
  playerId: number;
  name: string;
  avatar: string;
  compassion: number;
  courage: number;
  wisdom: number;
  ambition: number;
  principle: number;
  completedQuests: number[];
  archetypeEarned?: string | null;
}

export interface PlayerProfile {
  playerId: number;
  name: string;
  avatar: string;
  compassion: number;
  courage: number;
  wisdom: number;
  ambition: number;
  principle: number;
  completedQuests: number[];
  archetypeEarned?: string | null;
  previousArchetypes?: string[];
}

export interface UpdateStatsRequest {
  compassionDelta: number;
  courageDelta: number;
  wisdomDelta: number;
  ambitionDelta: number;
  principleDelta: number;
  questId: number;
  archetypeEarned?: string | null;
}

export interface UpdateAvatarRequest {
  avatar: string;
}

export interface ErrorResponse {
  message: string;
}