# Pyt_volshebnika/packages/api-zod/src/generated/types/playerProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wizard Path Game API
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerProfile {
  playerId: number;
  name: string;
  avatar: string;
  compassion: number;
  courage: number;
  wisdom: number;
  ambition: number;
  principle: number;
  completedQuests: number[];
  archetypeEarned?: string | null;
  previousArchetypes?: string[];
}