# Pyt_volshebnika/packages/api-zod/src/generated/types/updateStatsRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wizard Path Game API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateStatsRequest {
  compassionDelta: number;
  courageDelta: number;
  wisdomDelta: number;
  ambitionDelta: number;
  principleDelta: number;
  questId: number;
  archetypeEarned?: string | null;
}